@@ -1,16 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="12.0.1" />
-	<PackageReference Include="DocuSign.eSign.dll" Version="6.2.0" />
+    <PackageReference Include="AutoMapper" Version="13.0.1" />
+	<PackageReference Include="Azure.Storage.Blobs" Version="12.21.2" />
+	<PackageReference Include="DocuSign.eSign.dll" Version="7.0.0" />
 	<PackageReference Include="Expo.Server.SDK" Version="1.0.2" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.2" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.8" />
   </ItemGroup>
 
   <ItemGroup>
