@@ -1,8 +1,6 @@
 ﻿using Microsoft.Extensions.Options;
-using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
 using System.Net.Http.Headers;
-using System.Net.Http;
 using System.Text;
 using Traki.Domain.Services.Notifications;
 
@@ -32,7 +30,7 @@ namespace Traki.Infrastructure.Services
             var content = new StringContent(json, Encoding.UTF8, "application/json");
 
             // expo endpoint is not secured unless push security is enabled
-            if (!_notificationSettings.AccessToken.IsNullOrEmpty())
+            if (!string.IsNullOrEmpty(_notificationSettings.AccessToken))
             {
                 _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", _notificationSettings.AccessToken);
             }
